backend/requirements.txt
autopep8==1.6.0
black==22.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
ecdsa==0.17.0
finnhub-python==2.4.13
Flask==2.1.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mock-open==1.4.0
mypy-extensions==0.4.3
numpy==1.22.4
pandas==1.4.2
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.5.2
pyasn1==0.4.8
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.37
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
